requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
autopep8==1.4.4
blis==0.2.4
certifi==2019.6.16
cffi==1.12.3
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cymem==2.0.2
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-easy-select2==1.5.6
django-eventstream==2.5.0
django-filter==2.1.0
django-fluent-contents==2.0.6
django-fluent-utils==2.0.1
django-grip==2.0.0
django-invitations==1.9.2
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-parler==1.9.2
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-socketio==0.3.9
django-tag-parser==3.1
django-template-analyzer==1.6.1
djangorestframework==3.9.4
fr-core-news-md==2.1.0
future==0.17.1
gevent==1.4.0
gevent-socketio==0.2.1
gevent-websocket==0.10.1
greenlet==0.4.15
gripcontrol==3.3.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
joblib==0.13.2
jsonfield==2.0.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.4.4
numpy==1.16.4
Pillow==6.1.0
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.3
pubcontrol==2.4.2
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
spacy==2.1.6
sphinx-me==0.3
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.8
tqdm==4.32.2
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
wasabi==0.2.2
webencodings==0.5.1
Werkzeug==0.15.5
wrapt==1.11.2
zope.interface==4.6.0
